+++ src/main/java/com/example/demo/service/ValidationService.java
@@ -4,6 +4,6 @@ import com.example.demo.entity.ValidationEntity;
 
 public interface ValidationService{
     ValidationEntity postData(ValidationEntity validate);
-    
-     ValidationEntity getDataById(Integer id);
+
+     ValidationEntity getDataById(Long id);
 }
